+++ enum/src/main/java/org/grnet/status/enums/TenantJobEvent.java
@@ -1,5 +1,7 @@
 package org.grnet.status.enums;
 
+import io.quarkus.oidc.Tenant;
+
 import java.util.Arrays;
 import java.util.Optional;
 import java.util.Set;
@@ -26,7 +28,8 @@ public enum TenantJobEvent {
     INIT_MONITORING_BOX("init_monitoring_box", EventMode.AUTO, true, Set.of(TenantJobProperty.TENANT_ID, TenantJobProperty.TENANT_NAME)), // no properties for now
 
     INIT_POEM("init_poem", EventMode.AUTO, true, Set.of(TenantJobProperty.TENANT_ID, TenantJobProperty.TENANT_NAME)),
-    INIT_ARCHIVER("init_archiver", EventMode.AUTO, false, Set.of(TenantJobProperty.TENANT_ID, TenantJobProperty.TENANT_NAME));
+    INIT_ARCHIVER("init_archiver", EventMode.AUTO, false, Set.of(TenantJobProperty.TENANT_ID, TenantJobProperty.TENANT_NAME)),
+    INIT_PERFORMANCE_DATA("init_performance_data",EventMode.AUTO, false,Set.of(TenantJobProperty.TENANT_ID, TenantJobProperty.TENANT_NAME, TenantJobProperty.PERFORMANCE_DATA));
 
     private final String key;
     private final EventMode mode;