+++ CHANGELOG.md
@@ -68,6 +68,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#235](https://github.com/ARGOeu/argo-mon-api/pull/235) ARGO-5888 argo-mon-api: when a new tenant is created please in INIT_COMPUTE_ENGINE event specify if performance:true/false (according to what the user has defined in the tenant)- #235
 - [#236](https://github.com/ARGOeu/argo-mon-api/pull/236) ARGO-5862 Publicly Expose Status Timelines endpoints for tenant's reports
 - [#238](https://github.com/ARGOeu/argo-mon-api/pull/238) ARGO-5870 Add Service-Specific Monitoring Metrics Endpoints
+- [#241](https://github.com/ARGOeu/argo-mon-api/pull/241) ARGO-5880 create a INIT_PERFORMANCE_DATA event
 
 ### Fix